Handle external call updates during ringing in InCallWakeLockController.
In b/215559767, we changed the logic so that, for an HFP call, we firtly
mark them as a regular call, so that telecom stack will keep ringing.
And only if the user decided to pick up the call, we mark the external
call property depending on where this call is picked up. However,
telecom stack didn't consider that the call's external state would be
changed during ringing phase, so:

1. onCallAdded - The call is not an external call, so the telecom stack
   holds the wakelock and keeps ringing.
2. onCallStateChanged - The call is picked up, but since now the call is
   marked as external. We are not releasing the wake lock.

This patch captures the external call state change, and trigger
handleWakeLock again.

Bug: 220283618
Test: atest TelecomUnitTests:com.android.server.telecom.tests.InCallWakeLockControllerTest
Change-Id: If7861a8cbcaf7fbbc37c71d63d5f378c4e2e80cd
(cherry picked from commit 259aa4ecd31699de293cbc39981894f4f148a67b)