update when getDefaultDialerComponent sends a crash notif.
A recent change was made to InCallController#getDefaultDialerComponent
that sends a crash notification when the Default Dialer does not
implement an InCallService.  However, role requirements are only
enforced for apps targeting T.  Therefore, the change needs to be
stricter and an additional check has been added to ensure only
apps targeting T and above will receive the notification.

bug: 218903401
Test: 1 new unit test, 1 updated
Change-Id: I98df1e9a97b0184b924505ce3db46b8c689b96d2