fix log spewing in Telecom
The log lines "Could not get Analytics CallInfo" and
"EventRecord added as Call" were appearing repeatedly and at an absurd
magnitude (hundreds of thousands) in ATP test runs (log spewing).

It looks like the EventListener & SessionListener
were being re-initialized on multiple occasions causing the lines
to echo off all the listeners. Additionaly, for every new init, a new
listener was added.

To prevent this issue from happening, a bool check was placed on the
listeners to ensure they are initialized once.

bug: 224442099
Test: 1 new testing class LogUtilsTest & 1 unit test (testLogUtilsIsNotReInitialized)
Change-Id: I886774cf7ea75abfe02f757f495afb719f7c7dca