Removed gated notif. + refactor ICC#gDDC
It was noted that the argument Binder.getCallingUid to
InCallController#getDefaultDialerComponent was incorrect
leading to the crash notification to be shown on every
target sdk.

However,when running tests manually with the correct argument
(DefaultPhoneAppName), on test apps targeting T,the RoleManager
would kick the test DefaultPhoneApp off of the Role#Dialer.  This
made the crash notification useless since the if clause would never
be entered.

Therefore, the crash notification code has been removed and a
refactor was done on the getDefaultDialer method to make the
logic more clear.

bug: 226186102
Test: 2 modified, 1 removed
Change-Id: Ib356e42b1747fc6e06bd0f216b88ef96be2ef283