BluetoothLeAudio: Fix missing group id for device.
Telecom has registered callbacks for BluetoothLeAudio and in this way it
tracs devices being added/removed to/from group.
However, it could happen that callback is registered after device has
been added to the group. In such a case, Telecom will not be aware about
the group for the device. This patch fixes it.

Issue was visible when LeAudio device is already bonded and Bluetooth is
enabled.

Bug: 150670922
Bug: 227562754
Test: atest TelecomUnitTests

Change-Id: I81d13f2c629eb774cc277692da9e9f1bb31a0779