Remove usage of deprecated setting VIBRATE_WHEN_RINGING
Remove usage in Telecom service of the deprecated setting key
VIBRATE_WHEN_RINGING in favor of the existing RING_VIBRATION_INTENSITY.

The "vibrate first then ring gradually" setting configured by the
ramping ringer flag in AudioManager will only be applied if the ringtone
vibrations are enabled, which matches the user expectation of the
Settings in the updated screen.

If the ring vibrations are disabled, or the phone is in silent mode,
then the ramping ringer will not be applied and the ringtone will start
right away.

Fix: 229178667
Test: manual
Change-Id: I7ffc5b60fcc57b4f8535e185f248ab4e5079032c