Ensure correct audio mode is used for conferences.
Conferences don't have an ability to indicate they're using VOIP audio
mode.  As a consequence, CallAudioManager can end up setting the audio
mode back to MODE_INCALL when holding and holding a conference call where
the children are all VOIP calls.  To fix this, add ability to get the
voip mode from the children if available.

Also added some more logging for VOIP audio mode and changes to this
attribute on a call.

Test: Manual on network; make voip conference and verify that the audio
works after hold/unhold operation and verify that MODE_IN_COMMUNICATION
is set correctly.
Test: Add unit test to ensure voip audio mode is detected for a conference
with VOIP children.
Fixes: 230536808
Change-Id: Idd3faa5cb8d78b886e7c87901692a3b0b1bbe10b