Check CALL_PHONE and MANAGE_OWN_CALLS permission in placeCall
Staticaly check CALL_PHONE and MANAGE_OWN_CALLS depending on the
PhoneAccountHandle package name passed in instead of taking into
account whether or not the PhoneAccountHandle is associated with
a registered self-managed ConnectionService. This makes sure
we do not leak Telecom state with respect to which PhoneAccounts
are registered back to the caller.

Bug: 231988638
Test: atest TelecomUnitTests
Change-Id: Ie5eb859f79df22fc4710a0a745e8b6bd9b48aeb9