tracked calls now recieve the AR change for active call
WA reported a bug where a self managed call was placed via earpiece,
placed another MO PSTN call on speaker was made in between, and when coming
back to the WA call, the audio route was still earpiece and not speaker.

This change updates all tracked calls with the audio route change of the
active foreground call.

Bug: 233930427
Test: 2 manual,
	a. using telecom's self-managed test app
	b. using WA (ConnectionService version)  w/ whitelisted numbers
      1 unit test, atest CallAudioRouteStateMachineTest#
                         testTrackedCallsReceiveAudioRouteChange
Change-Id: I5362381bb8b230a2cf7aaa550940ced53e8d5c44