prohibit toggling setAudioModeIsVoip(false) for SM connection
For context, there was a report of a VoIP app experiencing audio
issues in call when switching between a self-managed and PSTN call.
The root cause was from calling setAudioModeIsVoip(false)
on a self-managed connection.

To prevent misuse of the API, prevent toggling of
setAudioModeIsVoIP(false) on a self-managed connection.

Additionally, 2 CTS tests and 1 CtsVerifier test are put in place to
assert there isnt manipulation of this API.

Test: 2 cts,
      1 CtsVerifier called "Telecom Call Audio test"

bug: 236018173

Change-Id: Idf014cd87119c2ed7902af0e460f60d030b92502