enforceCallingPackage for getCallStateUsingPackage
It was pointed out that getCallStateUsingPackage was vulnerable
to the exploit where the callingPackage can be replaced.  This had
been fixed once before but was reverted because the change was
causing test failures.

Turns out the Android shell UID (2000 group) was not permitted
to call getCallStateUsingPackage.  This was causing a Security
Exception to be thrown and ultimately what caused the p0 test
failures.

To allow the enforceCallingPackage to be called within the
getCallStateUsingPackage, Process.SHELL_UID is now on the
telecom allow list.

bug: 236813210
Test: manual.  CTS test would be too fragile.
Change-Id: I8c7ba07240fee51312557d7c3411f7603aace613