Fix for incoming and outgoing call not starting as top app
By not starting the service as a top app it caused it to grab default
scheduling priority (non RT) for a good part of the initial flow until
an activity is created by dialer which adds signficant latency per the
linux scheduler behavior. By adding this flag we make the service
start as a top app and ensure RT scheduling is set when service is
bound.

Bug: 240343304
Test: Manual and trace analysis
Change-Id: I1ad4f4352ef99571227899b00b3ae1d7ffa219f9