Fix vibration issue in DND mode.
Currently if user turn on the vibrator or set a ringtone with haptic
channel, we only evaluate the ringer mode using
AudioManager#getRingerMode. While in DND mode, this value will always be
SILENT. For cases that some call can bypass the dnd check, we need
bypass the ringer mode check when we try to start the vibration.

Bug: 244887005
Test: TelecomUnitTestCases:RingerTest
Change-Id: Ic327d81e247c429a44f2aaec34e1897c00338427