Telecom Transactional APIs Implementation
Transactional APIs are defined as APIs that use
android.os.OutcomeReceivers.  The receivers are to be completed by
Telecom for CallControl opertaions and the Client for CallEventCallback
operations.  Doing so ensures the client and telecom are in sync with
call states and that operations can be completed on both ends.

bug: 249779561
Test: unit + CTS

Change-Id: Ib4cc7c7b05491e4f61c011a5d7af5bc24125d34d