Fix recording of permission grants
While granting permissions, one of the grants could fail (with exception)
This results in EmergencyCallHelper not recording the permission grant.
So in scenarios where FINE_LOCATION was granted, but BACKGROUND location
was not granted (due to exception) --- we never end up revoking FINE_LOCATION perm.

Added UT for EmergencyCallHelper

See: b/245279980#comment5 for issue logs
Bug: 209812164
Test: UT plus manual verification with dialer. on b/209812164#18
Change-Id: I02a1a19336c6ed6ead0cd60414a00ea671bf7f98