Fix completableFuture exception handling in call.
Ensured runtime exceptions that occur in the call path are not
swallowed. Added locations where exceptions are caught and handled. If a
runtime exceptions occurs while connecting an emergency call, added a
check that sets the current call to disconnected and removed to prevent future
calls from failing to connect.

Bug: 243811811
Test: Manual Test
Change-Id: I5e0cf1489dc2e93e81b260696619580245565fac