Expose telecom API for retrieving photo uri
This API is queried when the async contact query for telecom
is performed, however, it is not currently exposed to users.
However, dialer apps sometimes need this information so they
resort to doing the exact same query that telecom does on their
own, adding significant latency to their incoming/outgoing call
latencies.

Bug: 240341306
Test: Manual, verified the data is sent with logs.

Change-Id: I992cf8a46440124fb1bb66539319d50f6eb5fd22