Update RoleManagerAdapter to be Work-Profile aware
Currently, RoleManagerAdapter gets the default app names redirection and
screening by looking at the primary user. We also bind to the CallScreening service as the primary user. Instead, we should be grabbing the appropriate user from the call (namely, from the account handle).

There is also an issue placing MO calls from work sims which is tied to
the call redirection flow. The phone account being used for the call is
checked against the CallManager's current user, which is always the
system. This results in calls ending up in the disconnected state.

Bug: 256241185
Bug: 263303152
Test: Unit tests, Manual
Change-Id: I7bbbe6e447dad3640f83e0de3bd6b48ae66b0a85