fix TelecomManager#getters for transactional accounts
Currently, if an application registers accounts with
CAPABILITY_SUPPORTS_TRANSACTIONAL_OPERATIONS and tries to fetch
the account via a TelecomManager#getter, the account will not be
returned.

The bug is a result of the check in
PhoneAccountRegistrar#resolveComponent that ensures there is a valid
ConnectionService for the application fetching the accounts.

In U, applications that utilize accounts with
CAPABILITY_SUPPORTS_TRANSACTIONAL_OPERATIONS do not need to create a
ConnectionService.  Therefore, this check should be bypassed for such
accounts.

Fixes: 266873410
Test: testFetchingTransactionalAccounts + manual
Change-Id: I7c4ba664230547bc7145c79d874bec3507101feb