Fix issue where Telecom ANRs due to call into AudioManager.
The audiomanager calls in this case are not related to anything Telecom
is doing, so we can move them onto their own single thread handler for
async one-shot tasks.  The functionality looks at the current voice call
stream volume and increases it to default at the start of the call if
needed.

Test: Added new unit tests to cover this functionality.
Fixes: 268579695
Change-Id: Iaa35d3025583cfeb75e56d52847a7d78b8ee518f