Add Telecom events for call logging.
When we get bugs to do with writing data to the call log, we have no way
to definitively indicate whether or not a call was logged to the call log,
and if so what number/postdial digits/presentation was actually logged.

This change adds some basic dumpsys info to indicate when a call is or is
not logged to the call log.

Test: Manual test of logging messages in Telecom dumpsys
Bug: 260962050
Change-Id: Iba6ea39c12609a27668aef88d01945eaad5e23c3