Refactor dependency on AccessibilityManager.
The class is final, so the attempt to mock it in the Telecom unit tests
was resulting in Mockito just using the actual class.  This was causing
test flakes.

To fix this, added a new interface that abstracts out the few methods
from AM that our code actually uses.  We mock these in the unit tests and
rely on the real class in production.

Test: Ran all Telecom unit tests to verify no regressions.
Fixes: 270428514
Change-Id: I0e58896640e5ff5c9d7446665ff0ca42cf15597c