Move outbound audiomanager calls to async task executor.
The setSpeakerphoneOn method calls out via 2-way binders to audiomanager
and status bar manager.  Refactoring that code so that it can happen
on an async task executor.

There was already a async executor in CallsManager (Added recently) so this
was refactored out to a common spot so it could be reused in both these
spots.

This prevents Telecom from potentially being blocked on outgoing calls
to the audio framework or the status bar manager.

Test: Run Telecom unit tests and verify pass.
Fixes: 268442423
Change-Id: I040da18528bd5c5a8fd738e65b2ba7c2c2503c05