Dont flag disconnected calls that are created but never added.
Calls that are intentionally disconnected by Telecom before they are added to CallsManager are working as intended and should not trigger anomaly reports. This CL stops the CallANomalyWatchdog from tracking these calls. This CL also ensures CallsManager notifies the watchdog appropriately when a call of this type is marked as disconnected.

Bug: 269467189
Test: atest CallAnomalyWatchdogTest && atest CallsManagerTest
Change-Id: Id47166fc5031221cd9686d133e7cc9516cc3f03a