Change semantics of onCallCreated to be onStartCreateConnection.
onCallCreated encapsulated other things which happen prior to the call
being added to the CallAnomalyWatchDog.  For outgoing calls in particular,
this includes the NewOutgoingCallBroadcast being sent and the
CallRedirectionService being run.  Repurposed this call back to instead
happen just before CreateConnetionPRocessor runs and the ConnetionService
is asked to start creating a new connection.

Also renamed onCallCreatedButNeverAdded to onCreateConnectionFailed for
symmetry.

This will clear up the anomalies we're seeing which are a result of the
slow NewOutgoingCallBroadcast.  The CallRedirectionService already has its
own timeout so we don't need CallAnomalyWatchdog to enforce one, and the
NOCB is going to be none blocking in the future we can ignore those cases
for now.

Test: Run Telecom unit tests
Fixes: 270068100
Change-Id: I47c8d2b0da2aa1027ba2b6d734cbd4080c80c4a5