update the BIND_TELECOM_CONNECTION_SERVICE error msg
The error message in PhoneAccountRegistrar#registerPhoneAccount is
outdated with the new development in U. Starting in U, apps do not
need a ConnectionService in order to register a PhoneAccount. This
is due to the fact that there is a new way of managing calls via a
transactional way. Apps indicate this with the PhoneAccount capability
called CAPABILITY_SUPPORTS_TRANSACTIONAL_OPERATIONS.

bug: 263413874
Test: manual
Change-Id: I189af17db5fc7c664ec7d148882a3c02accda5b8