Remove BIND_TO_IN_CALL_EMERGENCY_ERROR_UUID anomaly report.
Some ICS may fail to bind and that is normal. This anomaly report has already been removed for non-emergency calls. This CL removes this anomaly report trigger for emergency calls as well as cleaning up the variables that were created for reporting both emergency and non-emergency calls for this type of anomaly report.

Test: Run unit tests InCallControllerTests
Bug: 275168097
Change-Id: I1eefca89792a8be8489e312641af6e865f102261