Refactor set/clear communication device logic for call audio.
Consolidates the AudioManager set/clear communication logic into a
single class. There are several issues that are being observed when
switching between different call audio routes and making sure the device
in the bt_stack corresponds to the device that has been requested for
communication. One example that was observed was when we switched from
speaker to an LE audio device: we end up setting the LE audio device for
communication before clearing the speaker device. As a result,
AudioManager ends up treating it as a request to clear the LE audio
device.

This CL ensures that each request to set the communication
device will always clear that same device before setting a new device
for communication by keeping track of a notion of a locally requested
device. Instead of relying on AudioManager#getCommunicationDevice to see
which device was requested for communication, a local reference to the
device is being tracked within Telecom. All requests to set/clear
communication devices will depend on the state of this device.

Fixes: 265611304
Test: Modifed existing unit tests
Test: New unit test to validate edge case in speaker to BT audio
Change-Id: I78769aff0d9001c93f092af253478ea21944df09