fix toggling hold for a single transactional call
I discovered in manual testing calls that were set to inactive (held)
were not able to be set active again. This was due to the fact that the
requestFocusTransaction asserts the focusManager has the desired new
active call as the focusCall.  FocusManager will not set the focusCall
to a held call which would error out the transaction.

If there are no active calls, then the requested new active call does
not need to wait for a hold on another call and the focusManager will
update after the call is marked as active.

Fixes: 266621488
Test: atest android.telecom.cts.TransactionalApisTest
                                       #testToggleActiveAndInactive
Change-Id: Ife3d3b3b73ca40237f9a0a22e7c369b2b63e2c3a