Increase transitory state timeout to 10 sec.
We have seen some transitory state timeouts when a call is waiting for
call filtering to happen.  Call filtering has a 5 second timeout, which
was the exact same length as this transitory timeout.  Increasing the
transitory timeout to 10 sec to account for the fact that call filtering
has its own timeout.

Test: Re-ran unit tests
Fixes: 277250068
Change-Id: Ic625c1af9c7418c29189044a70376e5425a76728