Fix secondary user unable to access in-call screen
We bind to the InCallService for the user placing the call (this is
derived from the phone account). This was changed from binding to
UserHandle.CURRENT to support the incoming work profile changes
that were added into U. Falling back on binding the service on
UserHandle.CURRENT in scenarios where the user placing the call is not
running under a managed profile (i.e. work) and is not the same as
UserHandle.CURRENT will resolve this issue.

From manual testing, I can see that work calls are still binding to the
work profile dialer. For the secondary user, I can see the call screen
pop up while the call is dialing and we can click into the notification.
The calls made on the personal profile are still working as intended.

Fixes: 276611583
Test: Manual (tested calls on admin/work/secondary user)
Change-Id: Ic4e071b1831c89d26811d66081ac2ae2aaa96d72