Prevent non-emergency calls for secondary users.
Secondary users are able to place calls to non-emergency numbers from
dialer even if the user does not have the option to place calls enabled.
The culprit is in UserCallIntentProcessor where we process the outgoing
call intent and bypass the outgoing call restriction check because the
IPC call for TM#placeCall is a local invocation. Removing the condition
in this CL as it is incorrect.

Fixes: 277702915
Test: Manual (Can see dialog that only ecalls are permitted)
Change-Id: I161ef4a690dc7cdce3491efb314c0d933512d37a