Make sure late connected non ui ICS are unbound.
When a non-ui ICS changed its component status to enabled and we did a
binding in response to the ACTION_PACKAGE_CHANGED broadcast we were not
adding the connections to the mSubConnetions in the tracking class.  As
a result we'd never unbind in the disconnect method since it wasn't
tracked.

Also updated InCallControllerTest to have a specific test for this and
to change the UID/UserHandle references in that file to make sense given
how UIDs and user handles work.  An app's uid is a multiple of 100000
where the uid / 100000 is the user handle.

Test: Aded new unit test for this specific case.
Bug: 278722218
Change-Id: I0d3af5a82501b93290eb1a114fd7459703701200