Dual sim MT emergency call issue
When dual sim is enabled and we have a sim assigned to a work profile as
well as the personal, MT emergency calls do not succeed. When the
Telephony preferred emergency account is different from the sim that is
receiving the call, there would not be any actively ringing call for
that sim. We should instead receive the call over the phone account of
the MT call.

Fixes: 279070364
Test: Unit
Test: Manual (tested dual sim scenario on work/profile)
Test: Manual (tested single sim cases - no regression observed)
Change-Id: I46993ebe9d0b2e989703a148fddbd98c1c23a694