Fix conference log history for work profiles.
The calling user in ConnectionServiceWrapper#addExistingConnection is
always user 0, even in the context of work calls. This is problematic as
we leverage the calling user to retrieve all the phone accounts that are
visible for that user. In order to support the notion of multi-user
awareness, we can validate that the calling user has granted the
INTERACT_ACROSS_USERS permission in order to allow access to all phone
accounts.

Fixes: 279974597
Test: Manual (Can see conference logs in telecom dumpsys)
Change-Id: I53686c7e3e6f73050474c28e86e86c34b3e91ff6