Fix issue with late binding to ICS.
If a call is added to InCallController and NO bindings take place, Telecom
would not track that call in its call id mapper.  If one of the disabled
ICS becomes enabled later, Telecom would skip binding because it assumes
it should only have bound if there is a call present.

To fix this, moved the "addCall" invocation to happen all the time in
onCallAdded; this ensures that regardless of whether we bind to an ICS we
still know about the call if a package enables later on in the call.

Test: Added failing unit test and verified that code change fixes it.
Test: Manual test by modifying Telecom so that it doesn't do an initial
binding; verified that I was able to repro the original bug and that the
change fixed it.
Fixes: 280309857

Change-Id: I8783b2ce5be4af9e24d6ffa4bca81b23a64e0700