disconnect or ignore onAnswer/onSetAnswer CCCs that are rejected
TL;DR:
- docs update for CallControlCallback rejection cases
- CTS coverage for onAnswer#(wasCompleted(false)) case
- CTS coverage for onSetActive#(wasCompleted(false)) case
- TSW fix

In most of the Transactional tests, it's expected that the client
always completed the CallControlCallback. In adding these tests I
noticed that the current logic will answer or set the call active if the
client rejects the onAnswer or onSetActive. In order to prevent this two
active call scenario, I've adjusted the appropriate code.

Fixes: 279034589
Test: CTS coverage for changes
Change-Id: I614ac31d53b4bc01b35f40de9dbc5a75933f09c5