Fix holding behavior across self-manged ConnectionServices.
- CallsManager#holdActiveCallForNewCall - make logs more concise.
- CallsManager#makeRoomForOutgoingCall - there is behavior which assumes
that if a new call is added by the SAME connection service that we will
let the ConnectionService worry about hold/unhold on its own.  That is a
legacy behavior from Telephony and really doesn't make sense for VOIP
apps.  For VOIP apps we SHOULD auto hold the ongoing call when a new
outgoing call is added.  This was not noticed in the past because the focus
tracker was doing this in some cases so it wasn't obvious.  However, even
then the signalling was not optimal as you could have an ACTIVE and
DIALING call, when realistically as soon as the new dialing call is added
it should hold the old active call.
- CSFocusManager - add local log of focus call to dumpsys to make tracking
focus issues easier in dumpsys.
- Updated the test app to include another connection service which was
useful for reproing this issue.

Test: Manual tests with test app; make calls across different cs in same
app.
Test: Added new unit test coverage for these cases.
Fixes: 279669173

Change-Id: If7d8e9f7280ce1d78b16a0d3b2f0b50b1063a4e6